requirements.txt
python-telegram-bot==20.7
gspread==5.7.2
google-auth==2.29.0
openai==1.14.2
python-dotenv==1.0.0
SQLAlchemy==2.0.41
pytz==2023.3
requests==2.32.0
aiohttp==3.10.11
apscheduler==3.10.4
